Polished the file and Folders
